se sube gestion de equipos

+++ ms-equipos/src/main/java/ms_equipos/ms_equipos/Repository/EquiposRepository.java
@@ -3,5 +3,5 @@ package ms_equipos.ms_equipos.Repository;
 import ms_equipos.ms_equipos.Entity.Equipos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EquiposRepository extends JpaRepository<Equipos,Integer> {
+public interface EquiposRepository extends JpaRepository<Equipos,Long> {
 }
